Move public API and token to token.ts
